Update Fold docs, move containers ops to streamly pkg
Move the container dependent code from streamly-core to streamly
Rename some functions.
Release some functions.
