+++ core/streamly-core.cabal
@@ -239,7 +239,6 @@ library
                      , Streamly.Internal.Control.ForkIO
                      , Streamly.Internal.Data.Cont
                      , Streamly.Internal.System.IO
-                     , Streamly.Internal.Data.IsMap
 
                      -- streamly-strict-data
                      , Streamly.Internal.Data.Tuple.Strict
@@ -397,6 +396,11 @@ library
                       , Streamly.Internal.Data.Stream.StreamDK.Type
 
     build-depends:
+                    -- streamly-base
+                    --
+                    -- These dependencies can be reversed if we want
+                    -- streamly-base to depend only on base.
+                    --
                     -- Core libraries shipped with ghc, the min and max
                     -- constraints of these libraries should match with
                     -- the GHC versions we support. This is to make sure that
@@ -404,18 +408,23 @@ library
                     -- depending on doctest is a common example) can
                     -- depend on streamly.
                        base              >= 4.12  && < 4.18
-                     , containers        >= 0.6.0 && < 0.7
                      , deepseq           >= 1.4.4 && < 1.5
-                     , directory         >= 1.3.3 && < 1.4
                      , exceptions        >= 0.8.0 && < 0.11
                      , ghc-prim          >= 0.5.3 && < 0.10
                      , mtl               >= 2.2.2 && < 2.4
                      , transformers      >= 0.5.5 && < 0.7
+
+                    -- streamly-unicode-core
                      , template-haskell  >= 2.14  && < 2.20
 
+                     -- streamly-filesystem-core
+                     , directory         >= 1.3.3 && < 1.4
+
+                     -- fusion-plugin
                      , fusion-plugin-types >= 0.1 && < 0.2
 
                      -- XXX to be removed
+                     , containers        >= 0.6.0 && < 0.7
                      , heaps             >= 0.3     && < 0.5
     if !flag(use-unliftio)
       build-depends:   monad-control     >= 1.0 && < 1.1
