+++ src/Streamly/Internal/Data/IsMap.hs
