+++ streamly.cabal
@@ -337,7 +337,8 @@ library
                      -- dependency order To view dependency graph:
                      -- graphmod | dot -Tps > deps.ps
 
-                       Streamly.Internal.Data.IsMap.HashMap
+                       Streamly.Internal.Data.IsMap
+                     , Streamly.Internal.Data.IsMap.HashMap
 
                      -- streamly-concurrent
                      , Streamly.Internal.Control.Concurrent
@@ -352,6 +353,7 @@ library
                      , Streamly.Internal.Data.Fold.Concurrent.Channel.Type
                      , Streamly.Internal.Data.Fold.Concurrent.Channel
                      , Streamly.Internal.Data.Fold.Concurrent
+                     , Streamly.Internal.Data.Fold.Extra
 
                      , Streamly.Internal.Data.Stream.Concurrent.Channel.Type
                      , Streamly.Internal.Data.Stream.Concurrent.Channel.Dispatcher
@@ -364,6 +366,7 @@ library
                      , Streamly.Internal.Data.Stream.Zip.Concurrent
                      , Streamly.Internal.Data.Stream.Time
                      , Streamly.Internal.Data.Stream.Exception.Lifted
+                     , Streamly.Internal.Data.Stream.Extra
 
                      -- streamly-unicode (depends on unicode-data)
                      , Streamly.Internal.Unicode.Utf8
